changed the types of the entries for the color sensor methods

+++ src/shared_gui_delegate_on_robot.py
@@ -52,19 +52,19 @@ class Receiver(object):
 
     def go_straight_until_intensity_is_less_than(self, intensity_entry, speed_entry):
         print('Got go_straight_until_intensity_is_less_than', intensity_entry)
-        self.robot.drive_system.go_straight_until_intensity_is_less_than(float(intensity_entry), float(speed_entry))
+        self.robot.drive_system.go_straight_until_intensity_is_less_than(float(intensity_entry), int(speed_entry))
 
     def go_straight_until_intensity_is_greater_than(self, intensity_entry, speed_entry):
         print('Got go_straight_until_intensity_is_greater_than', intensity_entry)
-        self.robot.drive_system.go_straight_until_intensity_is_greater_than(float(intensity_entry), float(speed_entry))
+        self.robot.drive_system.go_straight_until_intensity_is_greater_than(float(intensity_entry), int(speed_entry))
 
     def go_straight_until_color_is(self, color_entry, speed_entry):
         print('Got go_straight_until_color_is', color_entry)
-        self.robot.drive_system.go_straight_until_color_is(color_entry, float(speed_entry))
+        self.robot.drive_system.go_straight_until_color_is(color_entry, int(speed_entry))
 
     def go_straight_until_color_is_not(self, color_entry, speed_entry):
         print('Got go_straight_until_color_is_not', color_entry)
-        self.robot.drive_system.go_straight_until_color_is_not(color_entry, float(speed_entry))
+        self.robot.drive_system.go_straight_until_color_is_not(color_entry, int(speed_entry))
 
     def display_camera_data(self):
         print("Got display_camera_data")
